+++ apps/demo/emails/magic-links/slack-confirm.tsx
@@ -14,7 +14,7 @@ import {
 } from '@react-email/components';
 
 interface SlackConfirmEmailProps {
-  validationCode?: string;
+  validationCode: string;
 }
 
 const baseUrl = process.env.VERCEL_URL
@@ -156,7 +156,7 @@ export const SlackConfirmEmail = ({
 
 SlackConfirmEmail.PreviewProps = {
   validationCode: 'DJZ-TLX',
-} as SlackConfirmEmailProps;
+};
 
 export default SlackConfirmEmail;