+++ packages/react-email/src/app/preview/[...slug]/preview.tsx
@@ -5,15 +5,22 @@ import { useState } from 'react';
 import { flushSync } from 'react-dom';
 import { Toaster } from 'sonner';
 import { useDebouncedCallback } from 'use-debounce';
-import type { EmailRenderingResult } from '../../../actions/render-email-by-path';
+import type { ControlsResult } from '../../../actions/get-email-controls';
+import {
+  type EmailRenderingResult,
+  renderEmailByPath,
+} from '../../../actions/render-email-by-path';
+import { updatePreviewProps } from '../../../actions/update-preview-props';
 import { CodeContainer } from '../../../components/code-container';
+import { PreviewPropControls } from '../../../components/preview-prop-controls';
 import {
   ResizableWarpper,
   makeIframeDocumentBubbleEvents,
 } from '../../../components/resizable-wrapper';
-import { Shell } from '../../../components/shell';
+import { Shell, ShellContent } from '../../../components/shell';
 import { Tooltip } from '../../../components/tooltip';
 import { useClampedState } from '../../../hooks/use-clamped-state';
+import { useEmailControls } from '../../../hooks/use-email-controls';
 import { useEmailRenderingResult } from '../../../hooks/use-email-rendering-result';
 import { useHotreload } from '../../../hooks/use-hot-reload';
 import { useRenderingMetadata } from '../../../hooks/use-rendering-metadata';
@@ -23,6 +30,10 @@ interface PreviewProps {
   slug: string;
   emailPath: string;
   pathSeparator: string;
+
+  previewProps: Record<string, unknown>;
+
+  serverControlsResult: ControlsResult;
   serverRenderingResult: EmailRenderingResult;
 }
 
@@ -30,6 +41,10 @@ const Preview = ({
   slug,
   emailPath,
   pathSeparator,
+
+  previewProps: initialPreviewProps,
+
+  serverControlsResult,
   serverRenderingResult,
 }: PreviewProps) => {
   const router = useRouter();
@@ -39,9 +54,18 @@ const Preview = ({
   const activeView = searchParams.get('view') ?? 'preview';
   const activeLang = searchParams.get('lang') ?? 'jsx';
 
-  const renderingResult = useEmailRenderingResult(
+  const [previewProps, setPreviewProps] =
+    useState<Record<string, unknown>>(initialPreviewProps);
+
+  const [renderingResult, setRenderingResult] = useEmailRenderingResult(
     emailPath,
+    previewProps,
     serverRenderingResult,
+  );
+  const previewPropsControls = useEmailControls(
+    emailPath,
+    serverControlsResult,
+    renderingResult,
   );
 
   const renderedEmailMetadata = useRenderingMetadata(
@@ -80,7 +104,18 @@ const Preview = ({
     router.push(`${pathname}?${params.toString()}`);
   };
 
-  const hasNoErrors = typeof renderedEmailMetadata !== 'undefined';
+  const debouncedUpdatePreviewProps = useDebouncedCallback(
+    (newProps: Record<string, unknown>) => {
+      updatePreviewProps(slug, newProps).catch((exception) => {
+        throw new Error('Could not update cookie for preview props', {
+          cause: exception,
+        });
+      });
+    },
+    200,
+  );
+
+  const hasErrors = 'error' in renderingResult;
 
   const [maxWidth, setMaxWidth] = useState(Number.POSITIVE_INFINITY);
   const [maxHeight, setMaxHeight] = useState(Number.POSITIVE_INFINITY);
@@ -130,8 +165,8 @@ const Preview = ({
       viewWidth={width}
     >
       {/* This relative is so that when there is any error the user can still switch between emails */}
-      <div
-        className="relative flex h-full bg-gray-200 pb-8"
+      <ShellContent
+        className="relative flex h-full bg-gray-200"
         ref={(element) => {
           const observer = new ResizeObserver((entry) => {
             const [elementEntry] = entry;
@@ -150,11 +185,10 @@ const Preview = ({
           };
         }}
       >
-        {'error' in renderingResult ? (
-          <RenderingError error={renderingResult.error} />
-        ) : null}
+        {hasErrors ? <RenderingError error={renderingResult.error} /> : null}
 
-        {hasNoErrors ? (
+        {/* If this is undefined means that the initial server render of the email had errors */}
+        {typeof renderedEmailMetadata !== 'undefined' ? (
           <>
             {activeView === 'preview' && (
               <ResizableWarpper
@@ -224,7 +258,35 @@ const Preview = ({
         ) : null}
 
         <Toaster />
-      </div>
+      </ShellContent>
+
+      {previewPropsControls ? (
+        <PreviewPropControls
+          controls={previewPropsControls}
+          onValueChange={(key, newValue) => {
+            const newPreviewProps = { ...previewProps, [key]: newValue };
+            setPreviewProps(newPreviewProps);
+
+            debouncedUpdatePreviewProps(newPreviewProps);
+
+            renderEmailByPath(emailPath, newPreviewProps, {
+              invalidatingRenderingCache: true,
+            })
+              .then((newRenderingResult) => {
+                setRenderingResult(newRenderingResult);
+              })
+              .catch((exception) => {
+                throw new Error(
+                  'Could not render the email after changing the props',
+                  {
+                    cause: exception,
+                  },
+                );
+              });
+          }}
+          previewProps={previewProps}
+        />
+      ) : null}
     </Shell>
   );
 };